Array-based and Link-based stack in Java
This is an implementation for both array-based and link-based stack in
Java by using generic. Stack.java is the interface of stack and
StackIsEmpty is the exception when the stack is empty by extends the
exception superclass.  In LinkedStack, an inner class node is defined
point to the position in the linked list.